+++ apm-agent-core/src/main/java/co/elastic/apm/configuration/CoreConfiguration.java
@@ -171,7 +171,8 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
         .aliasKeys("disabled_instrumentations")
         .configurationCategory(CORE_CATEGORY)
         .description("A list of instrumentations which should be disabled.\n" +
-            "Valid options are `jdbc`, `servlet-api`, `servlet-api-async`, `spring-mvc` and `incubating`.\n" +
+            "Valid options are `jdbc`, `servlet-api`, `servlet-api-async`, `spring-mvc`, `http-client`, `apache-httpclient`," +
+            "`spring-resttemplate` and `incubating`.\n" +
             "If you want to try out incubating features,\n" +
             "set the value to an empty string.")
         .buildWithDefault(Collections.<String>singleton("incubating"));