+++ apm-agent-plugins/apm-spring-resttemplate-plugin/src/main/java/co/elastic/apm/restclient/package-info.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,25 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package co.elastic.apm.httpclient;
+@NonnullApi
+package co.elastic.apm.restclient;
 
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.HttpClients;
-import org.junit.Before;
-
-public class ApacheHttpClientInstrumentationTest extends AbstractHttpClientInstrumentationTest {
-
-    private HttpClient client;
-
-    @Before
-    public void setUp() {
-        client = HttpClients.createDefault();
-    }
-
-    @Override
-    protected void performGet(String path) throws Exception {
-        client.execute(new HttpGet(path)).getStatusLine().getStatusCode();
-    }
-
-}
+import co.elastic.apm.annotation.NonnullApi;